# apps/task/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/projects": {
      "post": {
        "operationId": "ProjectsController_create",
        "summary": "",
        "description": "Create a new Project. The user must have Create permission for Projects.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProjectDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        },
        "tags": ["projects"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "ProjectsController_findAll",
        "summary": "",
        "description": "Get a list of all Projects. The user must have permission to Read all Projects.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Project" }
                }
              }
            }
          }
        },
        "tags": ["projects"],
        "security": [{ "bearer": [] }]
      }
    },
    "/projects/search": {
      "get": {
        "operationId": "ProjectsController_searchAll",
        "summary": "",
        "description": "Get a list of Projects on which the calling user have access to.\nThe user must have Read permission on the Projects.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Project" }
                }
              }
            }
          }
        },
        "tags": ["projects"],
        "security": [{ "bearer": [] }]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "summary": "",
        "description": "Get a Project. The calling user must have Read permission the Project.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        },
        "tags": ["projects"],
        "security": [{ "bearer": [] }]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "summary": "",
        "description": "Update a Project. The calling user must have Update permission the Project.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProjectDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        },
        "tags": ["projects"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "ProjectsController_remove",
        "summary": "",
        "description": "Delete a Project. The calling user must have Delete permission the Project.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["projects"],
        "security": [{ "bearer": [] }]
      }
    },
    "/projects/{id}/file-upload-signature": {
      "get": {
        "operationId": "ProjectsController_fileUploadSignature",
        "summary": "",
        "description": "Files are uploaded to cloudinary.com. For uploading a file, a signature is required.\nThis endpoint returns that signature which clients can use to upload files.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadSignatureResponseDto"
                }
              }
            }
          }
        },
        "tags": ["projects"],
        "security": [{ "bearer": [] }]
      }
    },
    "/projects/{id}/nda-url": {
      "get": {
        "operationId": "ProjectsController_findNdaUrl",
        "summary": "",
        "description": "Get Non Disclosure Agreement URL a Project.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindNdaUrlResponseDto"
                }
              }
            }
          }
        },
        "tags": ["projects"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "summary": "",
        "description": "Get a Task. The calling user must have Read permission the Task either explicitly or implicitly.\nExplicit Permission: E.g Given by a role e.g developer role\nImplicit Permission: The task criteria matches calling user profile, and calling user have permission to create Feedback.\nImplicit Permission: This user has created Feedback for this Task",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        },
        "tags": ["tasks"]
      },
      "patch": {
        "operationId": "TasksController_update",
        "summary": "",
        "description": "Update a Task. The calling user must have Update permission the Task.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "summary": "",
        "description": "Delete a Task. The calling user must have Delete permission the Task.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/tasks/{id}/task-requests": {
      "get": {
        "operationId": "TasksController_findTaskRequests",
        "summary": "",
        "description": "Get a list of all TaskRequests for a taskId.\nThe task must belong to the user.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TaskRequest" }
                }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/task-requests": {
      "post": {
        "operationId": "TaskRequestsController_create",
        "summary": "",
        "description": "Create a new TaskRequest. The user must have Create permission for TaskRequests.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTaskRequestDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskRequest" }
              }
            }
          }
        },
        "tags": ["task-requests"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "TaskRequestsController_findAll",
        "summary": "",
        "description": "Get a list of all TaskRequests. The user must have permission to Read all TaskRequests.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskRequest" }
              }
            }
          }
        },
        "tags": ["task-requests"],
        "security": [{ "bearer": [] }]
      }
    },
    "/task-requests/{id}": {
      "get": {
        "operationId": "TaskRequestsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskRequest" }
              }
            }
          }
        },
        "tags": ["task-requests"],
        "security": [{ "bearer": [] }]
      },
      "patch": {
        "operationId": "TaskRequestsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskRequestDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskRequest" }
              }
            }
          }
        },
        "tags": ["task-requests"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "TaskRequestsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["task-requests"],
        "security": [{ "bearer": [] }]
      }
    },
    "/task-requests/search-request/{taskId}": {
      "get": {
        "operationId": "TaskRequestsController_searchRequest",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "description": "Task id for which the user requested.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskRequest" }
              }
            }
          }
        },
        "tags": ["task-requests"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/feedbacks/tasks/{taskId}": {
      "get": {
        "operationId": "FeedbacksController_findFeedbacks",
        "summary": "",
        "description": "Get all Feedbacks for a Task. The user must have Read permission on the Task.",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "description": "The task id to list Feedbacks for",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Feedback" }
                }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": {
    "title": "Task",
    "description": "Task service is responsible for iterative validation of Prototypes. Tasks are created and Feedback is provided through this service.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "Bearer",
        "description": "Authentication is done by a signed JWT",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "ndaUrl": { "type": "string" },
          "ndaCloudId": { "type": "string" }
        },
        "required": ["title"]
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "status": { "enum": ["open", "closed"], "type": "string" },
          "userId": { "type": "string" },
          "ndaUrl": { "type": "string" },
          "ndaCloudId": { "type": "string" }
        },
        "required": [
          "id",
          "title",
          "description",
          "dateCreated",
          "status",
          "userId",
          "ndaUrl",
          "ndaCloudId"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["open", "closed"] },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "ndaUrl": { "type": "string" },
          "ndaCloudId": { "type": "string" }
        }
      },
      "FileUploadSignatureResponseDto": {
        "type": "object",
        "properties": {
          "apiKey": { "type": "string" },
          "uploadUrl": { "type": "string" },
          "signature": { "type": "string" },
          "timestamp": { "type": "number" },
          "folder": { "type": "string" }
        },
        "required": ["apiKey", "uploadUrl", "signature", "timestamp", "folder"]
      },
      "FindNdaUrlResponseDto": {
        "type": "object",
        "properties": { "ndaUrl": { "type": "string" } },
        "required": ["ndaUrl"]
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "description": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["text", "radio", "star-rating", "thumbs-rating"]
          },
          "order": { "type": "number", "minimum": 1 },
          "radioOptions": { "type": "array", "items": { "type": "string" } },
          "taskId": { "type": "string" }
        },
        "required": [
          "id",
          "description",
          "type",
          "order",
          "radioOptions",
          "taskId"
        ]
      },
...
      "Image": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "prototypeNumber": { "type": "number", "minimum": 1, "maximum": 2 },
          "taskId": { "type": "string" },
          "cloudId": { "type": "string" }
        },
        "required": [
          "id",
          "url",
          "dateCreated",
          "prototypeNumber",
          "taskId",
          "cloudId"
        ]
      },
      "Skill": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["id", "name"]
      },
      "Country": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "code": { "type": "string" }
        },
        "required": ["id", "name", "code"]
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "testType": { "enum": ["basic", "comparison"], "type": "string" },
          "prototypeFormat": {
            "enum": ["image", "iframe", "text"],
            "type": "string"
          },
          "status": { "enum": ["draft", "open", "closed"], "type": "string" },
          "accessType": {
            "enum": ["open", "nda", "request"],
            "type": "string"
          },
          "iframeUrl1": { "type": "string" },
          "iframeUrl2": { "type": "string" },
          "textualDescription1": { "type": "string" },
          "textualDescription2": { "type": "string" },
          "minAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "maxAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "minExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "maxExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "budget": { "type": "number" },
          "incentive": { "type": "number" },
          "userId": { "type": "string" },
          "projectId": { "type": "string" },
          "images": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Image" }
          },
          "skills": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Skill" }
          },
          "countries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Country" }
          },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" }
          },
          "feedbacks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Feedback" }
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "dateCreated",
          "testType",
          "prototypeFormat",
          "status",
          "accessType",
          "iframeUrl1",
          "iframeUrl2",
          "textualDescription1",
          "textualDescription2",
          "minAge",
          "maxAge",
          "minExperience",
          "maxExperience",
          "budget",
          "incentive",
          "userId",
          "projectId",
          "images",
          "skills",
          "countries",
          "questions",
          "feedbacks"
        ]
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "textAnswer": { "type": "string" },
          "radioAnswer": { "type": "string" },
          "starRatingAnswer": { "type": "number", "minimum": 1, "maximum": 5 },
          "thumbsRatingAnswer": { "type": "string", "enum": ["up", "down"] },
          "feedbackId": { "type": "string" },
          "questionId": { "type": "string" }
        },
        "required": [
          "id",
          "textAnswer",
          "radioAnswer",
          "starRatingAnswer",
          "thumbsRatingAnswer",
          "feedbackId",
          "questionId"
        ]
      },
      "Feedback": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "comment": { "type": "string" },
          "paymentStatus": {
            "enum": ["pending", "completed"],
            "type": "string"
          },
          "feedbackRating": { "type": "number" },
          "feedbackRatingComment": { "type": "string" },
          "taskRating": { "type": "number" },
          "taskRatingComment": { "type": "string" },
          "taskId": { "type": "string" },
          "task": { "$ref": "#/components/schemas/Task" },
          "userId": { "type": "string" },
          "answers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Answer" }
          }
        },
        "required": [
          "id",
          "dateCreated",
          "comment",
          "paymentStatus",
          "feedbackRating",
          "feedbackRatingComment",
          "taskRating",
          "taskRatingComment",
          "taskId",
          "task",
          "userId",
          "answers"
        ]
      },
...
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "testType": { "type": "string", "enum": ["basic", "comparison"] },
          "prototypeFormat": {
            "type": "string",
            "enum": ["image", "iframe", "text"]
          },
          "accessType": {
            "type": "string",
            "enum": ["open", "nda", "request"]
          },
          "iframeUrl1": { "type": "string" },
          "iframeUrl2": { "type": "string" },
          "textualDescription1": { "type": "string" },
          "textualDescription2": { "type": "string" },
          "minAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "maxAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "minExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "maxExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "budget": { "type": "number", "minimum": 0 },
          "incentive": { "type": "number", "minimum": 0 },
          "skills": { "type": "array", "items": { "type": "string" } },
          "countries": { "type": "array", "items": { "type": "string" } },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" }
          }
        }
      },
...
      "TaskRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "requestComment": { "type": "string" },
          "rejectionComment": { "type": "string" },
          "ndaUrl": { "type": "string" },
          "ndaCloudId": { "type": "string" },
          "requestStatus": {
            "enum": ["accepted", "requested", "rejected"],
            "type": "string"
          },
          "userId": { "type": "string" },
          "taskId": { "type": "string" }
        },
        "required": [
          "id",
          "dateCreated",
          "requestComment",
          "rejectionComment",
          "ndaUrl",
          "ndaCloudId",
          "requestStatus",
          "userId",
          "taskId"
        ]
      },
      "CreateTaskRequestDto": {
        "type": "object",
        "properties": {
          "requestComment": { "type": "string" },
          "rejectionComment": { "type": "string" },
          "ndaUrl": { "type": "string" },
          "requestStatus": {
            "type": "string",
            "enum": ["accepted", "requested", "rejected"]
          },
          "taskId": { "type": "string" },
          "ndaCloudId": { "type": "string" }
        },
        "required": ["taskId"]
      },
      "UpdateTaskRequestDto": {
        "type": "object",
        "properties": {
          "requestComment": { "type": "string" },
          "rejectionComment": { "type": "string" },
          "ndaUrl": { "type": "string" },
          "requestStatus": {
            "type": "string",
            "enum": ["accepted", "requested", "rejected"]
          },
          "ndaCloudId": { "type": "string" }
        }
      },
...
    }
  }
}